requirement.txt
tensorflow
transformers
datasets
pandas
sentence-transformers
spacy
flask-ngrok
numpy
opencv-python
pytesseract
pyspellchecker